Task 18

+++ core-service/src/main/java/ru/kolobkevic/market/core/controllers/ProductController.java
@@ -7,12 +7,13 @@ import org.springframework.validation.ObjectError;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import ru.kolobkevic.market.api.exceptions.ResourceNotFoundException;
-import ru.kolobkevic.market.core.dtos.ProductDto;
+import ru.kolobkevic.market.api.dtos.ProductDto;
 import ru.kolobkevic.market.core.converters.ProductConverter;
 import ru.kolobkevic.market.core.exceptions.DataValidationException;
 import ru.kolobkevic.market.core.entities.Product;
 import ru.kolobkevic.market.core.services.ProductService;
 
+import java.math.BigDecimal;
 import java.util.stream.Collectors;
 
 
@@ -25,8 +26,8 @@ public class ProductController {
 
     @GetMapping
     public Page<ProductDto> findAll(@RequestParam(defaultValue = "1", name = "p") int pageIndex,
-                                    @RequestParam(required = false, name = "minPrice") Double minPrice,
-                                    @RequestParam(required = false, name = "maxPrice") Double maxPrice) {
+                                    @RequestParam(required = false, name = "minPrice") BigDecimal minPrice,
+                                    @RequestParam(required = false, name = "maxPrice") BigDecimal maxPrice) {
         int pageSize = 10;
         if (pageIndex < 1) {
             pageIndex = 1;
