+++ core-service/src/main/java/ru/kolobkevic/market/core/entities/Product.java
@@ -6,6 +6,7 @@ import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Entity
@@ -19,7 +20,7 @@ public class Product {
     private Long id;
 
     @Column(name = "price")
-    private Double price;
+    private BigDecimal price;
 
     @Column(name = "title")
     private String title;
@@ -32,7 +33,7 @@ public class Product {
     @Column(name = "updated_at")
     private LocalDateTime updatedAt;
 
-    public Product(Long id, Double price, String title) {
+    public Product(Long id, BigDecimal price, String title) {
         this.id = id;
         this.price = price;
         this.title = title;
